Rocket now requires puzzle
A NPC appears when you buy fuel for the rocket this NPC will tell you that you need to solve a puzzle. Then it will bring you to the puzzle and you will need to complete it. For right now it doesn't do anything but in the future it will open up the rocket for you to explore future planets.